requirements.txt
langchain-openai
langchain
langchain-community
langgraph
python-dotenv
faiss-cpu
pypdf
fastapi
pydantic
uvicorn
